move resize_image to its own function
